Conflicts:
	charon/backends/ec2.py
	charon/deployment.py

+++ charon/backends/ec2.py
@@ -34,13 +34,15 @@ class EC2Definition(MachineDefinition):
         self.zone = x.find("attr[@name='zone']/string").get("value")
         self.controller = x.find("attr[@name='controller']/string").get("value")
         self.ami = x.find("attr[@name='ami']/string").get("value")
-        if self.ami == "": raise Exception("no AMI defined for EC2 machine ‘{0}’".format(self.name))
+        if self.ami == "":
+            raise Exception("no AMI defined for EC2 machine ‘{0}’".format(self.name))
         self.instance_type = x.find("attr[@name='instanceType']/string").get("value")
         self.key_pair = x.find("attr[@name='keyPair']/string").get("value")
         self.private_key = x.find("attr[@name='privateKey']/string").get("value")
         self.security_groups = [e.get("value") for e in x.findall("attr[@name='securityGroups']/list/string")]
         self.instance_profile = x.find("attr[@name='instanceProfile']/string").get("value")
         self.tags = {k.get("name"): k.find("string").get("value") for k in x.findall("attr[@name='tags']/attrs/attr")}
+
         def f(xml):
             return {'disk': xml.find("attrs/attr[@name='disk']/string").get("value"),
                     'size': int(xml.find("attrs/attr[@name='size']/int").get("value")),
@@ -49,6 +51,7 @@ class EC2Definition(MachineDefinition):
                     'deleteOnTermination': xml.find("attrs/attr[@name='deleteOnTermination']/bool").get("value") == "true",
                     'encrypt': xml.find("attrs/attr[@name='encrypt']/bool").get("value") == "true",
                     'passphrase': xml.find("attrs/attr[@name='passphrase']/string").get("value")}
+
         self.block_device_mapping = {_xvd_to_sd(k.get("name")): f(k) for k in x.findall("attr[@name='blockDeviceMapping']/attrs/attr")}
         self.elastic_ipv4 = x.find("attr[@name='elasticIPv4']/string").get("value")
 
@@ -69,14 +72,14 @@ class EC2State(MachineState):
     def get_type(cls):
         return "ec2"
 
-    state = charon.util.attr_property("state", MachineState.MISSING, int) # override
+    state = charon.util.attr_property("state", MachineState.MISSING, int)  # override
     public_ipv4 = charon.util.attr_property("publicIpv4", None)
     private_ipv4 = charon.util.attr_property("privateIpv4", None)
     elastic_ipv4 = charon.util.attr_property("ec2.elasticIpv4", None)
     access_key_id = charon.util.attr_property("ec2.accessKeyId", None)
     region = charon.util.attr_property("ec2.region", None)
     zone = charon.util.attr_property("ec2.zone", None)
-    controller = charon.util.attr_property("ec2.controller", None) # FIXME: not used
+    controller = charon.util.attr_property("ec2.controller", None)  # FIXME: not used
     ami = charon.util.attr_property("ec2.ami", None)
     instance_type = charon.util.attr_property("ec2.instanceType", None)
     key_pair = charon.util.attr_property("ec2.keyPair", None)
@@ -121,12 +124,10 @@ class EC2State(MachineState):
             self.dns_hostname = None
             self.dns_ttl = None
 
-
     def get_ssh_name(self):
         if not self.public_ipv4:
             raise Exception("EC2 machine ‘{0}’ does not have a public IPv4 address (yet)".format(self.name))
         return self.public_ipv4
-
 
     def get_private_key_file(self):
         if self.private_key_file: return self.private_key_file
@@ -143,7 +144,6 @@ class EC2State(MachineState):
         file = self.get_private_key_file()
         return ["-i", file] if file else []
 
-
     def get_physical_spec(self, machines):
         lines = ['    require = [ <nixos/modules/virtualisation/amazon-config.nix> ];']
 
@@ -153,7 +153,6 @@ class EC2State(MachineState):
                              .format(_sd_to_xvd(k), v['generatedKey']))
 
         return lines
-
 
     def get_keys(self):
         keys = MachineState.get_keys(self)
@@ -199,23 +198,23 @@ class EC2State(MachineState):
 
 
     def connect_route53(self):
-        if self._conn_route53: return
+        if self._conn_route53:
+            return
 
         # Get the secret access key from the environment or from ~/.ec2-keys.
         (access_key_id, secret_access_key) = charon.ec2_utils.fetch_aws_secret_key(self.route53_access_key_id)
 
         self._conn_route53 = boto.connect_route53(access_key_id, secret_access_key)
 
-
     def _get_instance_by_id(self, instance_id, allow_missing=False):
         """Get instance object by instance id."""
         self.connect()
         reservations = self._conn.get_all_instances([instance_id])
         if len(reservations) == 0:
-            if allow_missing: return None
+            if allow_missing:
+                return None
             raise Exception("EC2 instance ‘{0}’ disappeared!".format(instance_id))
         return reservations[0].instances[0]
-
 
     def _get_volume_by_id(self, volume_id):
         """Get instance object by instance id."""
@@ -225,7 +224,6 @@ class EC2State(MachineState):
             raise Exception("unable to find volume ‘{0}’".format(volume_id))
         return volumes[0]
 
-
     def _get_snapshot_by_id(self, snapshot_id):
         """Get snapshot object by instance id."""
         self.connect()
@@ -233,7 +231,6 @@ class EC2State(MachineState):
         if len(snapshots) != 1:
             raise Exception("unable to find snapshot ‘{0}’".format(snapshot_id))
         return snapshots[0]
-
 
     def _wait_for_ip(self, instance):
         self.log_start("waiting for IP address... ".format(self.name))
@@ -243,7 +240,8 @@ class EC2State(MachineState):
             self.log_continue("({0}) ".format(instance.state))
             if instance.state not in {"pending", "running", "scheduling", "launching", "stopped"}:
                 raise Exception("EC2 instance ‘{0}’ failed to start (state is ‘{1}’)".format(self.vm_id, instance.state))
-            if instance.ip_address: break
+            if instance.ip_address:
+                break
             time.sleep(3)
 
         self.log_end("{0} / {1}".format(instance.ip_address, instance.private_ip_address))
@@ -254,10 +252,8 @@ class EC2State(MachineState):
         self.public_ipv4 = instance.ip_address
         self.ssh_pinged = False
 
-
     def _booted_from_ebs(self):
         return self.root_device_type == "ebs"
-
 
     def update_block_device_mapping(self, k, v):
         x = self.block_device_mapping
@@ -266,7 +262,6 @@ class EC2State(MachineState):
         else:
             x[k] = v
         self.block_device_mapping = x
-
 
     def get_backups(self):
         self.connect()
@@ -295,6 +290,24 @@ class EC2State(MachineState):
                 backups[b_id]['info'] = info
         return backups
 
+    def remove_backup(self, backup_id):
+        self.log('removing backup {0}'.format(backup_id))
+        self.connect()
+        _backups = self.backups
+        if not backup_id in _backups.keys():
+            self.warn('backup {0} not found, skipping'.format(backup_id))
+        for dev, snapshot_id in _backups[backup_id].items():
+            snapshot = None
+            try:
+                snapshot = self._get_snapshot_by_id(snapshot_id)
+            except:
+                self.warn('Snapshot {0} not found, skipping.'.format(snapshot_id))
+            if not snapshot is None:
+                self.log('removing snapshot {0}'.format(snapshot_id))
+                snapshot.delete()
+
+        _backups.pop(backup_id)
+        self.backups = _backups
 
     def backup(self, backup_id):
         self.connect()
@@ -334,21 +347,22 @@ class EC2State(MachineState):
             # wait for available
             while True:
                 sys.stderr.write("[{0}] ".format(volume.status))
-                if volume.status == "available": break
+                if volume.status == "available":
+                    break
                 time.sleep(3)
                 volume.update()
             sys.stderr.write("\n")
 
             self.log("attaching volume ‘{0}’ to ‘{1}’".format(new_volume.id, self.name))
             new_volume.attach(self.vm_id, k)
-            self.block_device_mapping[k]['volumeId'] = new_volume.id # FIXME
-
+            self.block_device_mapping[k]['volumeId'] = new_volume.id  # FIXME
 
     def create(self, defn, check, allow_reboot):
         assert isinstance(defn, EC2Definition)
         assert defn.type == "ec2"
 
-        if self.state != self.UP: check = True
+        if self.state != self.UP:
+            check = True
 
         self.set_common_state(defn)
 
@@ -482,7 +496,8 @@ class EC2State(MachineState):
                 instance = self._get_instance_by_id(self.vm_id)
                 break
             except boto.exception.EC2ResponseError as e:
-                if e.error_code != "InvalidInstanceID.NotFound": raise
+                if e.error_code != "InvalidInstanceID.NotFound":
+                    raise
             self.log("EC2 instance ‘{0}’ not known yet, waiting...".format(self.vm_id))
             time.sleep(3)
 
@@ -520,7 +535,8 @@ class EC2State(MachineState):
                 self.log_start("waiting for machine to be in running state... ".format(self.name))
                 while True:
                     self.log_continue("({0}) ".format(instance.state))
-                    if instance.state == "running": break
+                    if instance.state == "running":
+                        break
                     if instance.state not in {"running", "pending"}:
                         raise Exception(
                             "EC2 instance ‘{0}’ failed to reach running state (state is ‘{1}’)"
@@ -541,7 +557,8 @@ class EC2State(MachineState):
                     instance.update()
                     while True:
                         self.log_continue("({0}) ".format(instance.ip_address))
-                        if instance.ip_address == defn.elastic_ipv4: break
+                        if instance.ip_address == defn.elastic_ipv4:
+                            break
                         time.sleep(3)
                         instance.update()
                     self.log_end("")
@@ -617,6 +634,7 @@ class EC2State(MachineState):
                 volume = self._get_volume_by_id(v['volumeId'])
                 if volume.volume_state() == "available":
                     self._conn.attach_volume(v['volumeId'], self.vm_id, k)
+
                 # Wait until the device is visible in the instance.
                 def check_dev():
                     res = self.run_command("test -e {0}".format(_sd_to_xvd(k)), check=False)
@@ -646,7 +664,7 @@ class EC2State(MachineState):
                              'attachment.device': k})
                 if len(volumes) != 1:
                     raise Exception("unable to find volume attached to ‘{0}’ on EC2 machine ‘{1}’".format(k, self.name))
-                v['volumeId'] = volumes[0].id # FIXME
+                v['volumeId'] = volumes[0].id  # FIXME
                 self.update_block_device_mapping(k, v)
 
         # Detach volumes that are no longer in the deployment spec.
@@ -711,7 +729,6 @@ class EC2State(MachineState):
         change.add_value(self.public_ipv4)
         changes.commit()
 
-
     def _delete_volume(self, volume_id):
         if not self.depl.confirm("are you sure you want to destroy EC2 volume ‘{0}’?".format(volume_id)):
             raise Exception("not destroying EC2 volume ‘{0}’".format(volume_id))
@@ -722,8 +739,8 @@ class EC2State(MachineState):
             volume.delete()
         except boto.exception.EC2ResponseError as e:
             # Ignore volumes that have disappeared already.
-            if e.error_code != "InvalidVolume.NotFound": raise
-
+            if e.error_code != "InvalidVolume.NotFound":
+                raise
 
     def destroy(self):
         if not self.vm_id: return True
@@ -731,15 +748,18 @@ class EC2State(MachineState):
 
         self.log_start("destroying EC2 machine... ".format(self.name))
 
-        instance = self._get_instance_by_id(self.vm_id)
-        instance.terminate()
+        instance = self._get_instance_by_id(self.vm_id, allow_missing=True)
 
-        # Wait until it's really terminated.
-        while True:
-            self.log_continue("({0}) ".format(instance.state))
-            if instance.state == "terminated": break
-            time.sleep(3)
-            instance.update()
+        if instance:
+            instance.terminate()
+
+            # Wait until it's really terminated.
+            while True:
+                self.log_continue("({0}) ".format(instance.state))
+                if instance.state == "terminated": break
+                time.sleep(3)
+                instance.update()
+
         self.log_end("")
 
         # Destroy volumes created for this instance.
@@ -759,14 +779,15 @@ class EC2State(MachineState):
         self.log_start("stopping EC2 machine... ".format(self.name))
 
         instance = self._get_instance_by_id(self.vm_id)
-        instance.stop() # no-op if the machine is already stopped
+        instance.stop()  # no-op if the machine is already stopped
 
         self.state = self.STOPPING
 
         # Wait until it's really stopped.
         while True:
             self.log_continue("({0}) ".format(instance.state))
-            if instance.state == "stopped": break
+            if instance.state == "stopped":
+                break
             if instance.state not in {"running", "stopping"}:
                 raise Exception(
                     "EC2 instance ‘{0}’ failed to stop (state is ‘{1}’)"
@@ -777,14 +798,14 @@ class EC2State(MachineState):
 
         self.state = self.STOPPED
 
-
     def start(self):
-        if not self._booted_from_ebs(): return
+        if not self._booted_from_ebs():
+            return
 
         self.log("starting EC2 machine".format(self.name))
 
         instance = self._get_instance_by_id(self.vm_id)
-        instance.start() # no-op if the machine is already started
+        instance.start()  # no-op if the machine is already started
 
         self.state = self.STARTING
 
@@ -801,9 +822,9 @@ class EC2State(MachineState):
 
         self.wait_for_ssh(check=True)
 
-
     def check(self):
-        if not self.vm_id: return
+        if not self.vm_id:
+            return
         self.connect()
         instance = self._get_instance_by_id(self.vm_id, allow_missing=True)
         old_state = self.state
@@ -823,7 +844,6 @@ class EC2State(MachineState):
         elif instance.state == "stopped":
             self.state = self.STOPPED
 
-
     def reboot(self):
         self.log("rebooting EC2 machine... ")
         instance = self._get_instance_by_id(self.vm_id)
@@ -833,6 +853,7 @@ class EC2State(MachineState):
 
 def _xvd_to_sd(dev):
     return dev.replace("/dev/xvd", "/dev/sd")
+
 
 def _sd_to_xvd(dev):
     return dev.replace("/dev/sd", "/dev/xvd")
